Создал компонент group-card-list, поправил указание css классов для card
